Split GroovyExtensionModuleTransformerTest
