+++ src/intiTest/kotlin/com/github/jengelman/gradle/plugins/shadow/transformers/TransformersTest.kt
@@ -5,7 +5,6 @@ import assertk.assertions.isEqualTo
 import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import com.github.jengelman.gradle.plugins.shadow.util.Issue
-import com.github.jengelman.gradle.plugins.shadow.util.doesNotContainEntries
 import com.github.jengelman.gradle.plugins.shadow.util.isRegular
 import kotlin.io.path.appendText
 import kotlin.io.path.writeText
@@ -154,135 +153,6 @@ class TransformersTest : BaseTransformerTest() {
     assertThat(mf2.mainAttributes.getValue("New-Entry")).isNull()
   }
 
-  @Test
-  fun groovyExtensionModuleTransformer() {
-    val one = buildJarOne {
-      insert(
-        ENTRY_SERVICE_EXTENSION_MODULE,
-        """
-          moduleName=foo
-          moduleVersion=1.0.5
-          extensionClasses=com.acme.foo.FooExtension,com.acme.foo.BarExtension
-          staticExtensionClasses=com.acme.foo.FooStaticExtension
-        """.trimIndent(),
-      )
-    }
-    val two = buildJarTwo {
-      insert(
-        ENTRY_SERVICE_EXTENSION_MODULE,
-        """
-          moduleName=bar
-          moduleVersion=2.3.5
-          extensionClasses=com.acme.bar.SomeExtension,com.acme.bar.AnotherExtension
-          staticExtensionClasses=com.acme.bar.SomeStaticExtension
-        """.trimIndent(),
-      )
-    }
-    projectScriptPath.appendText(
-      transform<GroovyExtensionModuleTransformer>(
-        shadowJarBlock = fromJar(one, two),
-      ),
-    )
-
-    run(shadowJarTask)
-
-    val props = outputShadowJar.getContent(ENTRY_SERVICE_EXTENSION_MODULE).toProperties()
-    assertThat(props.getProperty("moduleName")).isEqualTo("MergedByShadowJar")
-    assertThat(props.getProperty("moduleVersion")).isEqualTo("1.0.0")
-    assertThat(props.getProperty("extensionClasses"))
-      .isEqualTo("com.acme.foo.FooExtension,com.acme.foo.BarExtension,com.acme.bar.SomeExtension,com.acme.bar.AnotherExtension")
-    assertThat(props.getProperty("staticExtensionClasses"))
-      .isEqualTo("com.acme.foo.FooStaticExtension,com.acme.bar.SomeStaticExtension")
-  }
-
-  @Test
-  fun groovyExtensionModuleTransformerWorksForGroovy25Plus() {
-    val one = buildJarOne {
-      insert(
-        ENTRY_GROOVY_EXTENSION_MODULE,
-        """
-          moduleName=foo
-          moduleVersion=1.0.5
-          extensionClasses=com.acme.foo.FooExtension,com.acme.foo.BarExtension
-          staticExtensionClasses=com.acme.foo.FooStaticExtension
-        """.trimIndent(),
-      )
-    }
-    val two = buildJarTwo {
-      insert(
-        ENTRY_SERVICE_EXTENSION_MODULE,
-        """
-          moduleName=bar
-          moduleVersion=2.3.5
-          extensionClasses=com.acme.bar.SomeExtension,com.acme.bar.AnotherExtension
-          staticExtensionClasses=com.acme.bar.SomeStaticExtension
-        """.trimIndent(),
-      )
-    }
-    projectScriptPath.appendText(
-      transform<GroovyExtensionModuleTransformer>(
-        shadowJarBlock = fromJar(one, two),
-      ),
-    )
-
-    run(shadowJarTask)
-
-    val props = outputShadowJar.getContent(ENTRY_GROOVY_EXTENSION_MODULE).toProperties()
-    assertThat(props.getProperty("moduleName")).isEqualTo("MergedByShadowJar")
-    assertThat(props.getProperty("moduleVersion")).isEqualTo("1.0.0")
-    assertThat(props.getProperty("extensionClasses"))
-      .isEqualTo("com.acme.foo.FooExtension,com.acme.foo.BarExtension,com.acme.bar.SomeExtension,com.acme.bar.AnotherExtension")
-    assertThat(props.getProperty("staticExtensionClasses"))
-      .isEqualTo("com.acme.foo.FooStaticExtension,com.acme.bar.SomeStaticExtension")
-    assertThat(outputShadowJar).doesNotContainEntries(
-      ENTRY_SERVICE_EXTENSION_MODULE,
-    )
-  }
-
-  @Test
-  fun groovyExtensionModuleTransformerShortSyntax() {
-    val one = buildJarOne {
-      insert(
-        ENTRY_SERVICE_EXTENSION_MODULE,
-        """
-          moduleName=foo
-          moduleVersion=1.0.5
-          extensionClasses=com.acme.foo.FooExtension,com.acme.foo.BarExtension
-          staticExtensionClasses=com.acme.foo.FooStaticExtension
-        """.trimIndent(),
-      )
-    }
-    val two = buildJarTwo {
-      insert(
-        ENTRY_SERVICE_EXTENSION_MODULE,
-        """
-          moduleName=bar
-          moduleVersion=2.3.5
-          extensionClasses=com.acme.bar.SomeExtension,com.acme.bar.AnotherExtension
-          staticExtensionClasses=com.acme.bar.SomeStaticExtension
-        """.trimIndent(),
-      )
-    }
-    projectScriptPath.appendText(
-      """
-        $shadowJar {
-          ${fromJar(one, two)}
-          mergeGroovyExtensionModules()
-        }
-      """.trimIndent(),
-    )
-
-    run(shadowJarTask)
-
-    val props = outputShadowJar.getContent(ENTRY_SERVICE_EXTENSION_MODULE).toProperties()
-    assertThat(props.getProperty("moduleName")).isEqualTo("MergedByShadowJar")
-    assertThat(props.getProperty("moduleVersion")).isEqualTo("1.0.0")
-    assertThat(props.getProperty("extensionClasses"))
-      .isEqualTo("com.acme.foo.FooExtension,com.acme.foo.BarExtension,com.acme.bar.SomeExtension,com.acme.bar.AnotherExtension")
-    assertThat(props.getProperty("staticExtensionClasses"))
-      .isEqualTo("com.acme.foo.FooStaticExtension,com.acme.bar.SomeStaticExtension")
-  }
-
   @ParameterizedTest
   @MethodSource("transformerConfigurations")
   fun transformerShouldNotHaveDeprecatedBehaviours(pair: Pair<String, KClass<*>>) {
@@ -310,13 +180,11 @@ class TransformersTest : BaseTransformerTest() {
       "" to ApacheNoticeResourceTransformer::class,
       "" to ComponentsXmlResourceTransformer::class,
       "" to DontIncludeResourceTransformer::class,
-      "" to GroovyExtensionModuleTransformer::class,
       "{ resource.set(\"test.file\"); file.fileValue(file(\"test/some.file\")) }" to IncludeResourceTransformer::class,
       "" to Log4j2PluginsCacheFileTransformer::class,
       "" to ManifestAppenderTransformer::class,
       "" to ManifestResourceTransformer::class,
       "{ keyTransformer = { it.toLowerCase() } }" to PropertiesFileTransformer::class,
-      "" to XmlAppendingTransformer::class,
     )
   }
 }
